Unix client: wrappers for credentials structure access
This patch introduces and makes use of wrappers for access
to credentials structure members:
	cr_uid (afs_ucred_t *)
	cr_ruid(afs_ucred_t *)
	cr_gid (afs_ucred_t *)
	cr_rgid(afs_ucred_t *)
	cr_group_info(afs_ucred_t *)

Inline functions are also introduced to set values:
	set_cr_uid (afs_ucred_t *, uid_t)
	set_cr_ruid(afs_ucred_t *, uid_t)
	set_cr_gid (afs_ucred_t *, gid_t)
	set_cr_rgid(afs_ucred_t *, gid_t)
	set_cr_group_info(afs_ucred_t *, struct group_info *)

This will allow an architecture to make use of an alternate
structure to hold credentials.  In particular it will allow
the linux client to be modified to use the kernel credentials
structure directly instead of shadowing it into our own local
structure.

There should be no functional change.

Change-Id: I7739196a0691a9c30887043b085e4878e88780ae
Reviewed-on: http://gerrit.openafs.org/768
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>