Update afsd cache and firewall details
Cache parameters are discussed in two locations in the afsd man page,
and the first copy had not been updated for the new auto-tuning of
the chunk size and the stat parameter.  Fix both.

Note that the firewall requirements for klog only apply if you're using
kaserver and klog.  Kerberos v5 has its own requirements, but this is not
the place to talk about them.

Change-Id: I9cdaaa71351a64cecc1b6904efba87d4871d42fb
Reviewed-on: http://gerrit.openafs.org/798
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>