Shout less about system call hooks when using keyrings
When keyrings are in use, the user doesn't really care whether or
not we manage to hook the system call table, and the current message
has proved confusing, as it seems to indicate something has gone wrong.

Tone down the warning that's produced when keyrings are in use. This is
1.4.x only, as we're going to completely disable system call probes when
keyrings are in use in 1.5.x

FIXES 125215

Change-Id: I60bbe852d11f2f0c9546b30c877f5ff27d05e7fb
Reviewed-on: http://gerrit.openafs.org/1024
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>