viced: add then remove when changing addresses
When changing a host address, be sure to add the new interface
before removing the old to avoid having the host marked as
deleted if we removed the only valid address.  This change was
pulled out of commit ca613599a2537756462a420ae1a632747a433226
viced-host-uuid-and-addr-hashing-corrections-20090530 by
Jeffrey Altman.

LICENSE IPL10
FIXES 124634

Change-Id: Ie65e762c4be6205eb654566403f0d5088b7ebdee
Reviewed-on: http://gerrit.openafs.org/1026
Tested-by: Derrick Brashear <shadow@dementia.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>