up.c: remove unused variable pageSize
Variable pageSize is declared and set to the system page size
using getpagesize(), but the value is not used anywhere.

getpagesize() doesn't necessarily have a prototype declared in
unistd.h with recent glibc, which can generate a new warning and
trip up an enabled-checking build.

Instead of trying to put in a workaround, just remove the unused variable.

Change-Id: I6a9519eb2642bb0479edd1cf1195ce3f751c72bb
Reviewed-on: http://gerrit.openafs.org/1117
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>