Windows: drop cm_Freelance_Lock before call to cm_FreelanceAddMount
The cm_Freelance_Lock must not be held before calling
cm_FreelanceAddMount().  cm_InitLocalMountPoints() was violating
this requirement.   It is safe to drop the lock because the only
time the lock would be dropped is if the "Freelance" registry
key does not exist.  This should only be true during the initial
startup of the application the first time afsd_service.exe
executes on the machine.   At this point in time there is only
one thread that is executing that could be initializing the
Freelance mount point list.

LICENSE MIT

Change-Id: Ib49cf8bc830836c8e08446073e64ccf30086d354
Reviewed-on: http://gerrit.openafs.org/1164
Reviewed-by: Asanka Herath <asanka@secure-endpoints.com>
Tested-by: Asanka Herath <asanka@secure-endpoints.com>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
Tested-by: Jeffrey Altman <jaltman@openafs.org>